+++ src/main/java/chiochat/ChioChatException.java
@@ -1,3 +1,5 @@
+package chiochat;
+
 public class ChioChatException extends Exception {
     public ChioChatException(String message) {
         super(message);
