+++ src/main/java/chiochat/CommandManager.java
@@ -1,3 +1,4 @@
+package chiochat;
 import java.util.Map;
 import java.util.function.Consumer;
