+++ src/main/java/chiochat/Storage.java
@@ -1,3 +1,4 @@
+package chiochat;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
