+++ src/main/java/chiochat/Task.java
@@ -1,3 +1,4 @@
+package chiochat;
 public class Task {
     private final String description;
     private boolean isDone;
