--- text-ui-test/ip/src/main/java/data/TaskDB.txt
@@ -1,2 +1,0 @@
-T|0|hello world
-T|0|hello world
